% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_squares.R
\name{make_squares}
\alias{make_squares}
\title{try to create magic squares given a word list}
\usage{
make_squares(
  wordlist,
  quiet = TRUE,
  save_2row_hits = FALSE,
  save_3row_hits = FALSE
)
}
\arguments{
\item{wordlist}{vector of character string words (all the same number of characters, either all with 4 or all with 5)}

\item{quiet}{whether to print partial successes, attempts}

\item{save_2row_hits}{whether to save/return partial successes of just 2 rows that work}

\item{save_3row_hits}{whether to save/return partial successes of just 3 rows that work}
}
\value{
a list with counts of plausible combinations
  for 2,3,4,5 rows, and matrices of found partial or full squares
 The results include each valid table and the transposed version of it as well,
}
\description{
try to create magic squares given a word list
}
\examples{
 make_squares(c('AIR', 'ROE', 'END', 'ARE', 'ION', 'RED'))

 x = make_squares(  )
 x$found4
 transpose_each_square(x$found4)

}
