% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_squares.R
\name{print_squares}
\alias{print_squares}
\title{look at results to see if a given word or string query was among the squares found}
\usage{
print_squares(hitstable, query = "*", compact = TRUE)
}
\arguments{
\item{hitstable}{output of [make_squares()] function}

\item{query}{word or partial word, query for grepl, ignoring case}

\item{compact}{set FALSE to print with spaces between letters - easier to read the tables}
}
\value{
prints squares that had the query term
}
\description{
look at results to see if a given word or string query was among the squares found
}
